Additional checks for possible null options
Some locations did not have a location set when interacting
Ex. poachers, some gang fights(and some of those are bugged anyway) and
some post battle talks.
This would cause a null execption when trying to check if in tavern
New check early returns when location is null.
added characterobject being spoken to is the troop type of towns current
minor merc recruit